Updated Commands
Updated the welcomeoption and leaveoption commands so that they wait for messages instead of the info being put in the command. Updated the info commands and avatar command so that they now default to the channel the message was sent in/the user who sent the command if none was provided. Updated the ping command so it has an alias of "test".